+++ iota-client/src/api/address.rs
@@ -75,8 +75,8 @@ impl<'a> GetAddressesBuilder<'a> {
 
         let mut addresses = Vec::new();
         for i in range {
-            let address = generate_address(&seed, &mut path, i, false);
-            let internal_address = generate_address(&seed, &mut path, i, true);
+            let address = generate_address(&seed, &mut path, i, false)?;
+            let internal_address = generate_address(&seed, &mut path, i, true)?;
             let bech32_hrp = self._client.get_network_info().bech32_hrp;
             addresses.push((Bech32Address(address.to_bech32(&bech32_hrp)), false));
             addresses.push((Bech32Address(internal_address.to_bech32(&bech32_hrp)), true));
@@ -86,14 +86,11 @@ impl<'a> GetAddressesBuilder<'a> {
     }
 }
 
-fn generate_address(seed: &Seed, path: &mut BIP32Path, index: usize, internal: bool) -> Address {
+fn generate_address(seed: &Seed, path: &mut BIP32Path, index: usize, internal: bool) -> Result<Address> {
     path.push(internal as u32 + HARDEND);
     path.push(index as u32 + HARDEND);
 
-    let public_key = crate::secrets::Ed25519PrivateKey::generate_from_seed(seed, &path)
-        .expect("Invalid Seed & BIP32Path. Probably because the index of path is not hardened.")
-        .generate_public_key()
-        .to_compressed_bytes();
+    let public_key = seed.generate_private_key(path)?.public_key().to_compressed_bytes();
     // Hash the public key to get the address
     let mut hasher = VarBlake2b::new(32).unwrap();
     hasher.update(public_key);
@@ -105,7 +102,7 @@ fn generate_address(seed: &Seed, path: &mut BIP32Path, index: usize, internal: b
     path.pop();
     path.pop();
 
-    Address::Ed25519(Ed25519Address::new(result))
+    Ok(Address::Ed25519(Ed25519Address::new(result)))
 }
 
 /// Function to find the index and public or internal type of an Bech32 encoded address
